server/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
uvicorn
transformers